MSIM: Do not create a PhoneAccount for the slot without SIM card
When the slot has no SIM card, the dummy subId now is no longer
negative number, so the TelecomAccountRegistry will create a PhoneAccount
for the slot without SIM card too.
Check whether the slot has SIM card and it is provisoned. if yes, add a
PhoneAccount for the slot with SIM card.

Change-Id: Ibaed043f93a74d69e990594b4bb17d8e8c140a27
CRs-Fixed: 1015302 901710